Initialize ProjectData via config validator
