+++ esmvalcore/__init__.py
@@ -2,7 +2,6 @@
 import logging
 import os
 
-from ._projects import projects
 from ._session import session
 from ._version import __version__
 from .configuration import config
@@ -14,7 +13,6 @@ __all__ = [
     '__version__',
     'cmor',
     'config',
-    'projects',
     'session',
     'preprocessor',
 ]
